+++ src/components/index.js
@@ -1,0 +1,31 @@
+import MTableAction from './m-table-action';
+import MTableActions from './m-table-actions';
+import MTableBody from './m-table-body';
+import MTableBodyRow from './m-table-body-row';
+import MTableGroupbar from './m-table-groupbar';
+import MTableGroupRow from './m-table-group-row';
+import MTableCell from './m-table-cell';
+import MTableEditRow from './m-table-edit-row';
+import MTableEditField from './m-table-edit-field';
+import MTableFilterRow from './m-table-filter-row';
+import MTableHeader from './m-table-header';
+import MTablePagination from './m-table-pagination';
+import MTableSteppedPagination from './m-table-stepped-pagination';
+import MTableToolbar from './m-table-toolbar';
+
+export {
+  MTableAction,
+  MTableActions,
+  MTableBody,
+  MTableBodyRow,
+  MTableGroupbar,
+  MTableGroupRow,
+  MTableCell,
+  MTableEditRow,
+  MTableEditField,
+  MTableFilterRow,
+  MTableHeader,
+  MTablePagination,
+  MTableSteppedPagination,
+  MTableToolbar,
+};
