+++ src/components/m-table-edit-row.js
@@ -3,7 +3,7 @@ import { Checkbox, TableCell, TableRow, IconButton, Icon, Tooltip, Typography } 
 import PropTypes from 'prop-types';
 import * as React from 'react';
 import MTableCell from './m-table-cell';
-import { byString } from './utils';
+import { byString } from '../utils';
 /* eslint-enable no-unused-vars */
