+++ src/prop-types.js
@@ -1,0 +1,154 @@
+import PropTypes from 'prop-types';
+
+export const propTypes = {
+  actions: PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.func, PropTypes.shape({
+    icon: PropTypes.oneOfType([PropTypes.element, PropTypes.func, PropTypes.string]).isRequired,
+    isFreeAction: PropTypes.bool,
+    tooltip: PropTypes.string,
+    onClick: PropTypes.func.isRequired,
+    iconProps: PropTypes.object
+  })])),
+  columns: PropTypes.arrayOf(PropTypes.shape({
+    cellStyle: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+    currencySetting: PropTypes.shape({
+      locale: PropTypes.string,
+      currencyCode: PropTypes.string,
+      minimumFractionDigits: PropTypes.number,
+      maximumFractionDigits: PropTypes.number
+    }),
+    customFilterAndSearch: PropTypes.func,
+    customSort: PropTypes.func,
+    defaultFilter: PropTypes.any,
+    defaultSort: PropTypes.oneOf(['asc', 'desc']),
+    editComponent: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    emptyValue: PropTypes.oneOfType([PropTypes.string, PropTypes.node, PropTypes.func]),
+    export: PropTypes.bool,
+    field: PropTypes.string,
+    filtering: PropTypes.bool,
+    filterCellStyle: PropTypes.object,
+    grouping: PropTypes.bool,
+    headerStyle: PropTypes.object,
+    hidden: PropTypes.bool,
+    lookup: PropTypes.object,
+    editable: PropTypes.oneOf(['always', 'onUpdate', 'onAdd', 'never']),
+    removable: PropTypes.bool,
+    render: PropTypes.func,
+    searchable: PropTypes.bool,
+    sorting: PropTypes.bool,
+    title: PropTypes.oneOfType([PropTypes.element, PropTypes.string]),
+    type: PropTypes.oneOf(['string', 'boolean', 'numeric', 'date', 'datetime', 'time', 'currency'])
+  })).isRequired,
+  components: PropTypes.shape({
+    Action: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Actions: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Body: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Cell: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Container: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    EditField: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    EditRow: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    FilterRow: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Groupbar: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    GroupRow: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Header: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    OverlayLoading: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Pagination: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Row: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Toolbar: PropTypes.oneOfType([PropTypes.element, PropTypes.func])
+  }),
+  data: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.object), PropTypes.func]).isRequired,
+  editable: PropTypes.shape({
+    onRowAdd: PropTypes.func,
+    onRowUpdate: PropTypes.func,
+    onRowDelete: PropTypes.func
+  }),
+  detailPanel: PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.arrayOf(PropTypes.oneOfType([
+      PropTypes.func,
+      PropTypes.shape({
+        disabled: PropTypes.bool,
+        icon: PropTypes.oneOfType([PropTypes.element, PropTypes.func, PropTypes.string]),
+        openIcon: PropTypes.oneOfType([PropTypes.element, PropTypes.func, PropTypes.string]),
+        tooltip: PropTypes.string,
+        render: PropTypes.func.isRequired
+      })
+    ]))
+  ]),
+  icons: PropTypes.shape({
+    Add: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Check: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Clear: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Delete: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    DetailPanel: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Edit: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Export: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Filter: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    FirstPage: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    LastPage: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    NextPage: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    PreviousPage: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    ResetSearch: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    Search: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    SortArrow: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    ThirdStateCheck: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    ViewColumn: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+  }),
+  isLoading: PropTypes.bool,
+  title: PropTypes.oneOfType([PropTypes.element, PropTypes.string]),
+  options: PropTypes.shape({
+    actionsCellStyle: PropTypes.object,
+    actionsColumnIndex: PropTypes.number,
+    addRowPosition: PropTypes.oneOf(['first', 'last']),
+    columnsButton: PropTypes.bool,
+    defaultExpanded: PropTypes.bool,
+    debounceInterval: PropTypes.number,
+    detailPanelType: PropTypes.oneOf(['single', 'multiple']),
+    doubleHorizontalScroll: PropTypes.bool,
+    emptyRowsWhenPaging: PropTypes.bool,
+    exportAllData: PropTypes.bool,
+    exportButton: PropTypes.bool,
+    exportDelimiter: PropTypes.string,
+    exportFileName: PropTypes.string,
+    exportCsv: PropTypes.func,
+    filtering: PropTypes.bool,
+    filterCellStyle: PropTypes.object,
+    header: PropTypes.bool,
+    headerStyle: PropTypes.object,
+    initialPage: PropTypes.number,
+    maxBodyHeight: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    loadingType: PropTypes.oneOf(['overlay', 'linear']),
+    paging: PropTypes.bool,
+    pageSize: PropTypes.number,
+    pageSizeOptions: PropTypes.arrayOf(PropTypes.number),
+    paginationType: PropTypes.oneOf(['normal', 'stepped']),
+    rowStyle: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+    showEmptyDataSourceMessage: PropTypes.bool,
+    showSelectAllCheckbox: PropTypes.bool,
+    search: PropTypes.bool,
+    showTitle: PropTypes.bool,
+    toolbarButtonAlignment: PropTypes.oneOf(['left', 'right']),
+    searchFieldAlignment: PropTypes.oneOf(['left', 'right']),
+    searchFieldStyle: PropTypes.object,
+    selection: PropTypes.bool,
+    sorting: PropTypes.bool,
+    toolbar: PropTypes.bool,
+  }),
+  localization: PropTypes.shape({
+    grouping: PropTypes.shape({
+      groupedBy: PropTypes.string,
+      placeholder: PropTypes.string
+    }),
+    pagination: PropTypes.object,
+    toolbar: PropTypes.object,
+    header: PropTypes.object,
+    body: PropTypes.object
+  }),
+  initialFormData: PropTypes.object,
+  onSelectionChange: PropTypes.func,
+  onChangeRowsPerPage: PropTypes.func,
+  onChangePage: PropTypes.func,
+  onOrderChange: PropTypes.func,
+  onRowClick: PropTypes.func,
+  onTreeExpandChange: PropTypes.func,
+  tableRef: PropTypes.any
+};
